text sizes edited to correct form, phone number check is added to telefonniy spravochnik, and other bugs fixed.

+++ EDDS.Information/Helpfolder/GranitsiChastey.cs
@@ -197,11 +197,22 @@ namespace EDDS.Information.Helpfolder
                 {
                     Dialog4 dg4 = new Dialog4(DB, this, true, name, id);
                     dg4.ShowDialog(this);
+                    loadDatabase();
+                    UpdateRows();
                 }
 
                 if (e.ColumnIndex == 5)
-                    DB.ExecuteReader("DELETE FROM edds_sections WHERE id=" + id);
-                UpdateRows();
+                {
+                    DialogResult myResult;
+                    myResult = MessageBox.Show("Вы действительно хотите удалить\n" + dataGridView1.Rows[e.RowIndex].Cells[1].Value.ToString(), "Удалить", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+                    if (myResult == DialogResult.OK)
+                    {
+                        DB.ExecuteReader("DELETE FROM edds_sections WHERE id=" + id);
+                        MessageBox.Show("Удалено", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        loadDatabase();
+                        UpdateRows();
+                    }
+                }
             }
         }
 
@@ -289,8 +300,17 @@ namespace EDDS.Information.Helpfolder
                 }
 
                 if (e.ColumnIndex == 5)
-                    DB.ExecuteReader("DELETE FROM edds_sections WHERE id=" + id);
-                UpdateRows();
+                {
+                    DialogResult myResult;
+                    myResult = MessageBox.Show("Вы действительно хотите удалить\n" + dataGridView2.Rows[e.RowIndex].Cells[1].Value.ToString(), "Удалить", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+                    if (myResult == DialogResult.OK)
+                    {
+                        DB.ExecuteReader("DELETE FROM edds_sections WHERE id=" + id);
+                        MessageBox.Show("Удалено", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        loadDatabase();
+                        UpdateRows();
+                    }
+                }
             }
         }
 
@@ -458,7 +478,6 @@ namespace EDDS.Information.Helpfolder
         {
             UpdateRows();
         }
-
-       
+  
     }
 }
