+++ EDDS.Information/Helpfolder/KategoriVizova.cs
@@ -129,12 +129,23 @@ namespace EDDS.Information.Helpfolder
                 {
                     Dialog2 dg2 = new Dialog2("Обновление вызовы", DB, this, true, number, checkstate, id);
                     dg2.ShowDialog(this);
+                    loadDatabase();
+                    UpdateRows();
                 }
 
                 if (e.ColumnIndex == 4)
-                    DB.ExecuteReader("DELETE FROM edds_call_type WHERE id=" + id);
-                loadDatabase();
-                UpdateRows();
+                {
+                    DialogResult myResult;
+                    myResult = MessageBox.Show("Вы действительно хотите удалить\n" + dataGridView1.Rows[e.RowIndex].Cells[1].Value.ToString(), "Удалить", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+                    if (myResult == DialogResult.OK)
+                    {
+                        DB.ExecuteReader("DELETE FROM edds_call_type WHERE id=" + id);
+                        MessageBox.Show("Удалено", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        loadDatabase();
+                        UpdateRows();
+                    }
+                }   
+               
             }
         }
